// web/src/api/generated/models/ProfilePreferenceCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI ist der verbindliche Vertrag.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PreferenceCategory } from './PreferenceCategory';
import {
    PreferenceCategoryFromJSON,
    PreferenceCategoryFromJSONTyped,
    PreferenceCategoryToJSON,
    PreferenceCategoryToJSONTyped,
} from './PreferenceCategory';
import type { ProfileVisibility } from './ProfileVisibility';
import {
    ProfileVisibilityFromJSON,
    ProfileVisibilityFromJSONTyped,
    ProfileVisibilityToJSON,
    ProfileVisibilityToJSONTyped,
} from './ProfileVisibility';
import type { PreferenceSentiment } from './PreferenceSentiment';
import {
    PreferenceSentimentFromJSON,
    PreferenceSentimentFromJSONTyped,
    PreferenceSentimentToJSON,
    PreferenceSentimentToJSONTyped,
} from './PreferenceSentiment';

/**
 * 
 * @export
 * @interface ProfilePreferenceCreate
 */
export interface ProfilePreferenceCreate {
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceCreate
     */
    accountId: string;
    /**
     * 
     * @type {PreferenceCategory}
     * @memberof ProfilePreferenceCreate
     */
    category: PreferenceCategory;
    /**
     * 
     * @type {PreferenceSentiment}
     * @memberof ProfilePreferenceCreate
     */
    sentiment: PreferenceSentiment;
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceCreate
     */
    topic: string;
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceCreate
     */
    value: string;
    /**
     * 
     * @type {ProfileVisibility}
     * @memberof ProfilePreferenceCreate
     */
    visibility: ProfileVisibility;
}



/**
 * Check if a given object implements the ProfilePreferenceCreate interface.
 */
export function instanceOfProfilePreferenceCreate(value: object): value is ProfilePreferenceCreate {
    if (!('accountId' in value) || value['accountId'] === undefined) return false;
    if (!('category' in value) || value['category'] === undefined) return false;
    if (!('sentiment' in value) || value['sentiment'] === undefined) return false;
    if (!('topic' in value) || value['topic'] === undefined) return false;
    if (!('value' in value) || value['value'] === undefined) return false;
    if (!('visibility' in value) || value['visibility'] === undefined) return false;
    return true;
}

export function ProfilePreferenceCreateFromJSON(json: any): ProfilePreferenceCreate {
    return ProfilePreferenceCreateFromJSONTyped(json, false);
}

export function ProfilePreferenceCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfilePreferenceCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'accountId': json['accountId'],
        'category': PreferenceCategoryFromJSON(json['category']),
        'sentiment': PreferenceSentimentFromJSON(json['sentiment']),
        'topic': json['topic'],
        'value': json['value'],
        'visibility': ProfileVisibilityFromJSON(json['visibility']),
    };
}

export function ProfilePreferenceCreateToJSON(json: any): ProfilePreferenceCreate {
    return ProfilePreferenceCreateToJSONTyped(json, false);
}

export function ProfilePreferenceCreateToJSONTyped(value?: ProfilePreferenceCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'accountId': value['accountId'],
        'category': PreferenceCategoryToJSON(value['category']),
        'sentiment': PreferenceSentimentToJSON(value['sentiment']),
        'topic': value['topic'],
        'value': value['value'],
        'visibility': ProfileVisibilityToJSON(value['visibility']),
    };
}