// web/src/api/generated/models/ProfilePreferenceUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI ist der verbindliche Vertrag.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PreferenceCategory } from './PreferenceCategory';
import {
    PreferenceCategoryFromJSON,
    PreferenceCategoryFromJSONTyped,
    PreferenceCategoryToJSON,
    PreferenceCategoryToJSONTyped,
} from './PreferenceCategory';
import type { PreferenceSentiment } from './PreferenceSentiment';
import {
    PreferenceSentimentFromJSON,
    PreferenceSentimentFromJSONTyped,
    PreferenceSentimentToJSON,
    PreferenceSentimentToJSONTyped,
} from './PreferenceSentiment';

/**
 * 
 * @export
 * @interface ProfilePreferenceUpdate
 */
export interface ProfilePreferenceUpdate {
    /**
     * 
     * @type {PreferenceCategory}
     * @memberof ProfilePreferenceUpdate
     */
    category: PreferenceCategory;
    /**
     * 
     * @type {PreferenceSentiment}
     * @memberof ProfilePreferenceUpdate
     */
    sentiment: PreferenceSentiment;
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceUpdate
     */
    topic: string;
    /**
     * 
     * @type {string}
     * @memberof ProfilePreferenceUpdate
     */
    value: string;
}



/**
 * Check if a given object implements the ProfilePreferenceUpdate interface.
 */
export function instanceOfProfilePreferenceUpdate(value: object): value is ProfilePreferenceUpdate {
    if (!('category' in value) || value['category'] === undefined) return false;
    if (!('sentiment' in value) || value['sentiment'] === undefined) return false;
    if (!('topic' in value) || value['topic'] === undefined) return false;
    if (!('value' in value) || value['value'] === undefined) return false;
    return true;
}

export function ProfilePreferenceUpdateFromJSON(json: any): ProfilePreferenceUpdate {
    return ProfilePreferenceUpdateFromJSONTyped(json, false);
}

export function ProfilePreferenceUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfilePreferenceUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'category': PreferenceCategoryFromJSON(json['category']),
        'sentiment': PreferenceSentimentFromJSON(json['sentiment']),
        'topic': json['topic'],
        'value': json['value'],
    };
}

export function ProfilePreferenceUpdateToJSON(json: any): ProfilePreferenceUpdate {
    return ProfilePreferenceUpdateToJSONTyped(json, false);
}

export function ProfilePreferenceUpdateToJSONTyped(value?: ProfilePreferenceUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'category': PreferenceCategoryToJSON(value['category']),
        'sentiment': PreferenceSentimentToJSON(value['sentiment']),
        'topic': value['topic'],
        'value': value['value'],
    };
}